[testsuite] [riscv] xfail some [PR113281] tests
Some of the tests regressed with a fix for the vectorization of
shifts.  The riscv cost models need to be adjusted to avoid the
unprofitable optimization.  The failure of these tests has been known
since 2024-03-13, without a forthcoming fix, so I suggest we consider
it expected by now.  Adjust the tests to reflect that expectation.


for  gcc/testsuite/ChangeLog

	PR tree-optimization/113281
	* gcc.dg/vect/costmodel/riscv/rvv/pr113281-1.c: XFAIL.
	* gcc.dg/vect/costmodel/riscv/rvv/pr113281-2.c: Likewise.
	* gcc.dg/vect/costmodel/riscv/rvv/pr113281-5.c: Likewise.