[testsuite] [riscv] limit vwaddsub-1.c to rv64
The desired vw{add,sub}.wx instructions don't come up on rv32 for the
first two functions, we get v{add,sub}.vx instead.

I suppose this is an oversight, and something about the test is meant
for rv64 only, but the fact that the instruction is spelled out in the
intrinsic name and a different instruction is generated suggests
something may be wrong after all.


for  gcc/testsuite/ChangeLog

	* gcc.target/riscv/rvv/base/vwaddsub-1.c: Require rv64.