tailc: Use the IPA-VRP tail call hack even for pointers [PR119614]
As the first two testcases show, even with pointers IPA-VRP can optimize
return values from functions if they have singleton ranges into just the
exact value, so we need to virtually undo that for tail calls similarly
to integers and floats.  The third test just adds check that it works
even with floats (which it does).

2025-04-04  Jakub Jelinek  <jakub@redhat.com>

	PR tree-optimization/119614
	* tree-tailcall.cc (find_tail_calls): Handle also pointer types in the
	IPA-VRP workaround.

	* c-c++-common/pr119614-1.c: New test.
	* c-c++-common/pr119614-2.c: New test.
	* c-c++-common/pr119614-3.c: New test.