c++/modules: Propagate bits to DECL_MAYBE_DELETED dups [PR119462]
In the linked PR, we're importing over a DECL_MAYBE_DELETED decl with a
decl that has already been instantiated.  This patch ensures that the
needed bits are propagated across and that DECL_MAYBE_DELETED is cleared
from the existing decl, so that later synthesize_method doesn't crash
due to a definition unexpectedly already existing.

	PR c++/119462

gcc/cp/ChangeLog:

	* module.cc (trees_in::is_matching_decl): Propagate exception
	spec and constexpr to DECL_MAYBE_DELETED; clear if appropriate.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/noexcept-3_a.C: New test.
	* g++.dg/modules/noexcept-3_b.C: New test.
	* g++.dg/modules/noexcept-3_c.C: New test.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>
Reviewed-by: Jason Merrill <jason@redhat.com>