LoongArch: Add LoongArch architecture detection to __float128 support in libgfortran and libquadmath [PR119408].
In GCC14, LoongArch added __float128 as an alias for _Float128.
In commit r15-8962, support for q/Q suffixes for 128-bit floating point
numbers.  This will cause the compiler to automatically link libquadmath
when compiling Fortran programs.  But on LoongArch `long double` is
IEEE quad, so there is no need to implement libquadmath.
This causes link failure.

	PR target/119408

libgfortran/ChangeLog:

	* acinclude.m4: When checking for __float128 support, determine
	whether the current architecture is LoongArch.  If so, return false.
	* configure: Regenerate.

libquadmath/ChangeLog:

	* configure.ac: When checking for __float128 support, determine
	whether the current architecture is LoongArch.  If so, return false.
	* configure: Regenerate.

	Sigend-off-by: Xi Ruoyao <xry111@xry111.site>
	Sigend-off-by: Jakub Jelinek <jakub@redhat.com>