Ada: Fix wrong 'Access to aliased constrained array of controlled type
For technical reasons, the recently reimplemented finalization machinery
for controlled types requires arrays of controlled types to be allocated
with their bounds, including in the case where their nominal subtype is
constrained.  However, in this case, the type of 'Access for the arrays
is pointer-to-constrained-array and, therefore, its value must designate
the array itself and not the bounds.

gcc/ada/
	* gcc-interface/utils.cc (convert) <POINTER_TYPE>: Use fold_convert
	to convert between thin pointers.  If the source is a thin pointer
	with zero offset from the base and the target is a pointer to its
	array, displace the pointer after converting it.
	* gcc-interface/utils2.cc (build_unary_op) <ATTR_ADDR_EXPR>: Use
	fold_convert to convert the address before displacing it.